Move MousePointer to graphicsgeo
