+++ libs/graphicsgeo/src/main/kotlin/mono/graphics.geo/MousePointer.kt
@@ -1,0 +1,13 @@
+package mono.graphics.geo
+
+/**
+ * A sealed class which indicates mouse event pointer types.
+ */
+sealed class MousePointer {
+    object Idle : MousePointer()
+    data class Down(val point: Point) : MousePointer()
+    data class Move(val mouseDownPoint: Point, val point: Point) : MousePointer()
+    data class Up(val mouseDownPoint: Point, val point: Point) : MousePointer()
+    data class Click(val point: Point) : MousePointer()
+}
+
