+++ libs/htmlcanvas/src/main/kotlin/mono/html/canvas/mouse/MouseEventObserver.kt
@@ -1,5 +1,6 @@
 package mono.html.canvas.mouse
 
+import mono.graphics.geo.MousePointer
 import mono.graphics.geo.Point
 import mono.livedata.LiveData
 import mono.livedata.MutableLiveData
@@ -57,15 +58,4 @@ class MouseEventObserver(
     }
 
     private fun MouseEvent.toPoint(): Point = Point(toBoardLeft(clientX), toBoardTop(clientY))
-}
-
-/**
- * A sealed class which indicates mouse event pointer types.
- */
-sealed class MousePointer {
-    object Idle : MousePointer()
-    data class Down(val point: Point) : MousePointer()
-    data class Move(val mouseDownPoint: Point, val point: Point) : MousePointer()
-    data class Up(val mouseDownPoint: Point, val point: Point) : MousePointer()
-    data class Click(val point: Point) : MousePointer()
 }
